+++ api/snapshot.go
@@ -60,7 +60,7 @@ func (API *API) ListSnapshots(serverUUID string) ([]APISnapshot, *APIReturn, err
 	return snapshots, nil, nil
 }
 
-func (API *API) PostCreateSnapshot(serverUUID string) (*APISnapshot, *APIReturn, error) {
+func (API *API) CreateSnapshot(serverUUID string) (*APISnapshot, *APIReturn, error) {
 	rawData, apiReturn, err := API.SendRequestToAPI(HTTPPost, "/compute/servers/"+serverUUID+"/snapshots", nil)
 
 	// Communication error
@@ -80,4 +80,21 @@ func (API *API) PostCreateSnapshot(serverUUID string) (*APISnapshot, *APIReturn,
 		return nil, nil, err
 	}
 	return &snapshot, nil, nil
+}
+
+func (API *API) RestoreSnapshot(serverUUID, snapshotUUID string) (*APIReturn, error) {
+	// Send request
+	rawData, apiReturn, err := API.SendRequestToAPI(HTTPPut, "/compute/servers/"+
+		serverUUID+"/snapshots/"+snapshotUUID+"/restore", nil)
+
+	// Communication error
+	if err != nil {
+		return nil, err
+	}
+
+	// Unmarshal error
+	if apiReturn == nil {
+		return nil, errors.New(string(rawData))
+	}
+	return apiReturn, nil
 }